Trait embedded_nal_async::TcpConnect

pub trait TcpConnect {
    type Error: Error;
    type Connection<'a>: Read<Error = Self::Error> + Write<Error = Self::Error>
       where Self: 'a;

    // Required method
    async fn connect<'a>(
        &'a self,
        remote: SocketAddr,
    ) -> Result<Self::Connection<'a>, Self::Error>;
}

This trait is implemented by TCP/IP stacks. The trait allows the underlying driver to construct multiple connections that implement the I/O traits from embedded-io-async.

The associated connection type should close the connection when dropped.

Required Associated Types

type Error: Error

Error type returned on connect failure.

type Connection<'a>: Read<Error = Self::Error> + Write<Error = Self::Error> where Self: 'a

Type holding state of a TCP connection. Should close the connection when dropped.

Required Methods

async fn connect<'a>( &'a self, remote: SocketAddr, ) -> Result<Self::Connection<'a>, Self::Error>

Connect to the given remote host and port.

Returns Ok if the connection was successful.

Object Safety

This trait is not object safe.